fix settings.php for visualsvn configuration items cannot be saved. VisualSvnAuthPathsProvider: add $m_children as cache, $m_repo_name, implement lots functions.
confirm views and actions that use $appEngine->getAccessPathViewProvider() can use VisualSvnAuthPathsProvider
accesspathview.php only check the repo
repositorylist.php may keep, but changed to ignore accesspath check for efficiency
repositoryview.php only check the repo
actions/delete_repository.php skip remove_accesspaths with VisualSvnAuthPathsProvider

some views works well without changes:
accesspathslist.php
userview.php
groupview.php
permissionassign.php

some minor fixes:
AuthFileGroupAndPathProvider --> AuthFileGroupAndPathsProvider
IPathsEditProvider & IPathsViewProvider : isSplitByRepository --> hasChildren

+++ actions/delete_repository.php
@@ -66,11 +66,12 @@ else {
 			}
         	$appEngine->addMessage(tr("The repository %0 has been deleted.", array($oR->name)));
 
+			$accessPathViewProvider = $appEngine->getAccessPathViewProvider()
         	// Remove Access-Paths of the repository.
-        	if ($remove_accesspaths)
+        	if ($remove_accesspaths && !$accessPathViewProvider.hasChildren())
         	{
         		try {
-					$apList = $appEngine->getAccessPathViewProvider()->getPathsOfRepository($oR);
+					$apList = $accessPathViewProvider->getPathsOfRepository($oR);
 					foreach ($apList as $ap)
 					{
 						$appEngine->getAccessPathEditProvider()->deleteAccessPath($ap);
